extracted the console related classes
