+++ game.gemspec
@@ -1,13 +1,16 @@
 Gem::Specification.new do |s|
-    s.name        = 'multiple_play_game'
+    s.name        = 'game'
     s.version     = '1.0.0'
     s.date        = '2020-03-24'
     s.summary     = "TicTacToe Game"
     s.description = "A tictactoe game for console"
     s.authors     = ["Julius Ngwu"]
     s.email       = 'juliusczar.jc@gmail.com'
-    s.files         = Dir.chdir(File.expand_path('..', __FILE__)) do
-      `git ls-files -z`.split("\x0").reject { |f| f.match(%r{^(test|spec|features)/}) }
-    end
+    s.files        = ["lib/game.rb", 
+      "lib/tic_tac_toe_game/play.rb",
+      "lib/tic_tac_toe_game/move.rb",
+      "lib/tic_tac_toe_game/toggle.rb",
+      "lib/tic_tac_toe_game/language_settings.json"
+    ]
     s.license       = 'MIT'
   end
