+++ lib/console/continue_game.rb
