+++ lib/console/human_computer.rb
