+++ lib/console/launch_game.rb
