+++ lib/console/output.rb
@@ -1,5 +1,5 @@
 require "json"
-require "./lib/tic_tac_toe_game/language_option"
+require "./lib/console/language_option"
 
 class Output 
     def initialize(new_language_option = LanguageOption.new(JSON.parse(File.read("./lib/tic_tac_toe_game/language_settings.json"))))
